+++ plugins/Moodle/includes/ImportUsers.fnc.php
@@ -1,0 +1,515 @@
+<?php
+/**
+ * Import Moodle Users functions
+ *
+ * @package Moodle plugin
+ */
+
+/**
+ * List Moodle users
+ *
+ * @since 5.9
+ *
+ * @return array Empty if URL or Token not set or invalid, else users.
+ */
+function MoodleUsersList( $key, $value )
+{
+	require_once 'plugins/Moodle/client.php';
+
+	// Check Moodle URL if set + token set.
+	if ( ! MOODLE_URL
+		|| ! MOODLE_TOKEN )
+	{
+		return array();
+	}
+
+	$serverurl = MOODLE_URL . '/webservice/xmlrpc/server.php?wstoken=' . MOODLE_TOKEN;
+
+	if ( ! filter_var( $serverurl, FILTER_VALIDATE_URL ) )
+	{
+		return array();
+	}
+
+	// Check URL is responding with cURL.
+	$functionname = 'core_user_get_users';
+
+	// Dummy response function.
+	function core_user_get_users_response( $response )
+	{
+		// We had a response, return true so moodle_xmlrpc_call will return true.
+		return $response;
+	}
+
+	$criteria = array(
+		'key' => $key,
+		'value' => $value,
+	);
+
+	$object = array( 'criteria' => $criteria );
+
+	$users = moodle_xmlrpc_call( $functionname, $object );
+
+	return $users['users'];
+}
+
+/**
+ * Filter Moodle Users
+ * Users where confirmed=true, suspended=false, and id not exists in MOODLEXROSARIO table.
+ *
+ * @since 5.9
+ *
+ * @param  array $users Moodle Users.
+ *
+ * @return array        Filtered Moodle Users.
+ */
+function MoodleUsersFilter( $users )
+{
+	$filtered_users = array();
+
+	$moodlexrosario_user_ids_RET = DBGet( "SELECT MOODLE_ID
+		FROM MOODLEXROSARIO
+		WHERE " . DBEscapeIdentifier( 'COLUMN' ) . "='student_id'
+		OR " . DBEscapeIdentifier( 'COLUMN' ) . "='staff_id'", array(), array( 'MOODLE_ID' ) );
+
+	$moodlexrosario_user_ids = array_keys( $moodlexrosario_user_ids_RET );
+
+	foreach ( $users as $user )
+	{
+		if ( empty( $user['confirmed'] )
+			|| ! empty( $user['suspended'] )
+			|| in_array( $user['id'], $moodlexrosario_user_ids ) )
+		{
+			// User is not confirmed, or suspended, or already in RosarioSIS.
+			continue;
+		}
+
+		if ( $user['id'] == 1 )
+		{
+			// Guest User.
+			continue;
+		}
+
+		$filtered_users[ $user['id'] ] = $user;
+	}
+
+	return $filtered_users;
+}
+
+/**
+ * Make Moodle Users
+ * Format Users array for ListOutput.
+ *
+ * @since 5.9
+ *
+ * @param  array $users Filtered Moodle Users.
+ *
+ * @return array        Formatted Moodle Users.
+ */
+function MoodleUsersMake( $users )
+{
+	$formatted_users = array( 0 => array() );
+
+	foreach ( $users as $user )
+	{
+		if ( ! isset( $user['firstname'] ) )
+		{
+			// First and last name require 'moodle/site:viewfullnames' capabitility.
+			$names = explode( ' ', $user['fullname'] );
+
+			$user['firstname'] = $names[0];
+
+			$user['lastname'] = $names[1];
+		}
+
+		$formatted_users[] = array(
+			'CHECKBOX' => MakeChooseCheckbox( $user['id'] ),
+			'PROFILE' => MoodleUsersMakeProfile( $user['id'] ),
+			'FIRST_NAME' => MoodleUsersMakeName( $user['firstname'], $user['profileimageurl'] ),
+			'LAST_NAME' => $user['lastname'],
+			'EMAIL_ADDRESS' => ( empty( $user['email'] ) ? '' : $user['email'] ),
+			'USERNAME' => $user['username'],
+			'ID' => $user['id'],
+		);
+	}
+
+	unset( $formatted_users[0] );
+
+	return $formatted_users;
+}
+
+function MoodleUsersMakeProfile( $user_id )
+{
+	$profiles = array(
+		'student' => _( 'Student' ),
+		'teacher' => _( 'Teacher' ),
+		'parent' => _( 'Parent' ),
+		'admin' => _( 'Administrator' ),
+	);
+
+	$profile_select = SelectInput(
+		'student',
+		'values[PROFILE][' . $user_id . ']',
+		'',
+		$profiles,
+		false,
+		'',
+		false
+	);
+
+	return $profile_select;
+}
+
+function MoodleUsersMakeName( $name, $photo_url )
+{
+	require_once 'ProgramFunctions/TipMessage.fnc.php';
+
+	if ( ! $photo_url )
+	{
+		return $name;
+	}
+
+	return makeTipMessage(
+		'<img src="' . $photo_url . '" style="max-height: 150px;" />',
+		_( 'User Photo' ),
+		$name
+	);
+}
+
+
+function MoodleUsersStudentEnrollmentForm()
+{
+	$html = '<fieldset><legend>' . _( 'Enrollment' ) . '</legend><table class="width-100p">';
+
+	$gradelevels_RET = DBGet( DBQuery( "SELECT ID,TITLE
+		FROM SCHOOL_GRADELEVELS
+		WHERE SCHOOL_ID='" . UserSchool() . "'
+		ORDER BY SORT_ORDER" ) );
+
+	$options = array();
+
+	foreach ( (array) $gradelevels_RET as $gradelevel )
+	{
+		$options[ $gradelevel['ID'] ] = $gradelevel['TITLE'];
+	}
+
+	$html .= '<tr class="st"><td>' . SelectInput(
+		'',
+		'values[GRADE_ID]',
+		_( 'Grade Level' ),
+		$options,
+		false
+	) . '</td>';
+
+	$calendars_RET = DBGet( DBQuery( "SELECT CALENDAR_ID,DEFAULT_CALENDAR,TITLE
+		FROM ATTENDANCE_CALENDARS
+		WHERE SYEAR='" . UserSyear() . "'
+		AND SCHOOL_ID='" . UserSchool() . "'
+		ORDER BY DEFAULT_CALENDAR ASC" ) );
+
+	$options = array();
+
+	foreach ( (array) $calendars_RET as $calendar )
+	{
+		$options[ $calendar['CALENDAR_ID'] ] = $calendar['TITLE'];
+	}
+
+	$html .= '<td>' . SelectInput(
+		'',
+		'values[CALENDAR_ID]',
+		_( 'Calendar' ),
+		$options,
+		false
+	) . '</td>';
+
+	$schools_RET = DBGet( DBQuery( "SELECT ID,TITLE
+		FROM SCHOOLS
+		WHERE ID!='" . UserSchool() . "'
+		AND SYEAR='" . UserSyear() . "'" ) );
+
+	$options = array(
+		UserSchool() => _( 'Next grade at current school' ),
+		'0' => _( 'Retain' ),
+		'-1' => _( 'Do not enroll after this school year' ),
+	);
+
+	foreach ( (array) $schools_RET as $school )
+	{
+		$options[ $school['ID'] ] = $school['TITLE'];
+	}
+
+	$html .= '<td>' . SelectInput(
+		'',
+		'values[NEXT_SCHOOL]',
+		_( 'Rolling / Retention Options' ),
+		$options,
+		false
+	) . '</td></tr>';
+
+	$enrollment_codes_RET = DBGet( DBQuery( "SELECT ID,TITLE AS TITLE
+		FROM STUDENT_ENROLLMENT_CODES
+		WHERE SYEAR='" . UserSyear() . "'
+		AND TYPE='Add'
+		ORDER BY SORT_ORDER" ) );
+
+	$options = array();
+
+	foreach ( (array) $enrollment_codes_RET as $enrollment_code )
+	{
+		$options[ $enrollment_code['ID'] ] = $enrollment_code['TITLE'];
+	}
+
+	$default = DBGetOne( "SELECT min(SCHOOL_DATE) AS START_DATE
+		FROM ATTENDANCE_CALENDAR
+		WHERE SYEAR='" . UserSyear() . "'
+		AND SCHOOL_ID='" . UserSchool() . "'" );
+
+	if ( ! $default
+		|| DBDate() > $default )
+	{
+		$default = DBDate();
+	}
+
+	$html .= '<tr><td colspan="3">' . DateInput( $default, 'values[START_DATE]', '', false ) . ' - ' .
+		SelectInput(
+			'',
+			'values[ENROLLMENT_CODE]',
+			_( 'Attendance Start Date this School Year' ),
+			$options
+		) .	'</td></tr>';
+
+	$html .= '</table></fieldset>';
+
+	return $html;
+}
+
+
+function MoodleUserImportStudent( $user )
+{
+	global $error;
+
+	$username = $user['username'];
+
+	$email_field_key = Config( 'STUDENTS_EMAIL_FIELD' );
+
+	if ( $email_field_key !== 'USERNAME' )
+	{
+		$email_field_key = 'CUSTOM_' . Config( 'STUDENTS_EMAIL_FIELD' );
+	}
+	elseif ( ! empty( $user['email'] ) )
+	{
+		$username = $user['email'];
+	}
+
+	// Check username unicity.
+	$existing_username = DBGet( "SELECT 'exists'
+		FROM STAFF
+		WHERE USERNAME='" . $username . "'
+		AND SYEAR='" . UserSyear() . "'
+		UNION SELECT 'exists'
+		FROM STUDENTS
+		WHERE USERNAME='" . $username . "'
+		AND STUDENT_ID!='" . UserStudentID() . "'" );
+
+	if ( $existing_username )
+	{
+		$error[] = $username . ': ' . _( 'A user with that username already exists. Choose a different username and try again.' );
+
+		return 0;
+	}
+
+	do
+	{
+		$student_id = DBSeqNextID( 'students_student_id_seq' );
+	}
+	while ( DBGetOne( "SELECT STUDENT_ID
+		FROM STUDENTS
+		WHERE STUDENT_ID='" . $student_id . "'" ) );
+
+	if ( ! isset( $user['firstname'] ) )
+	{
+		// First and last name requires 'moodle/site:viewfullnames' capabitility.
+		$names = explode( ' ', $user['fullname'] );
+
+		$user['firstname'] = $names[0];
+
+		$user['lastname'] = $names[1];
+	}
+
+	$sql = "INSERT INTO STUDENTS ";
+	$fields = 'STUDENT_ID,LAST_NAME,FIRST_NAME,USERNAME';
+	$values = "'" . $student_id . "','" . $user['firstname'] . "','" . $user['lastname'] . "','" . $username . "'";
+
+	if ( $email_field_key !== 'USERNAME' )
+	{
+		$fields .= ',' . $email_field_key;
+		$values .= ",'" . $user['email'] . "'";
+	}
+
+	if ( ! empty( $_REQUEST['values']['PASSWORD_SET_USE_USERNAME'] ) )
+	{
+		$fields .= ',PASSWORD';
+		$values .= ",'" . encrypt_password( $username ) . "'";
+	}
+
+	$sql .= '(' . $fields . ') values(' . $values . ')';
+	DBQuery( $sql );
+
+	DBQuery( "INSERT INTO MOODLEXROSARIO (\"column\",rosario_id,moodle_id)
+		VALUES('student_id','" . $student_id . "'," . $user['id'] . ")" );
+
+	return $student_id;
+}
+
+function MoodleUserEnrollStudent( $student_id )
+{
+	$sql = "INSERT INTO STUDENT_ENROLLMENT ";
+
+	$fields = 'ID,SYEAR,SCHOOL_ID,STUDENT_ID,';
+
+	// Uses new serial ID sequence.
+	$values = db_seq_nextval( 'student_enrollment_id_seq' );
+
+	$values .= ",'" . UserSyear() . "','" . UserSchool() . "','" . $student_id . "',";
+
+	$fields .= 'START_DATE,GRADE_ID,ENROLLMENT_CODE,NEXT_SCHOOL,CALENDAR_ID';
+
+	$start_date = RequestedDate(
+		$_REQUEST['year_values']['START_DATE'],
+		$_REQUEST['month_values']['START_DATE'],
+		$_REQUEST['day_values']['START_DATE']
+	);
+
+	$values .= "'" . $start_date . "','" .
+		$_REQUEST['values']['GRADE_ID'] . "','" .
+		$_REQUEST['values']['ENROLLMENT_CODE'] . "','" .
+		$_REQUEST['values']['NEXT_SCHOOL'] . "','" .
+		$_REQUEST['values']['CALENDAR_ID'] . "'";
+
+	$sql .= '(' . $fields . ') values(' . $values . ');';
+	DBQuery( $sql );
+}
+
+
+function MoodleUserImportUser( $user, $profile )
+{
+	global $error;
+
+	$username = $user['username'];
+
+	// Check username unicity.
+	$existing_username = DBGet( "SELECT 'exists'
+		FROM STAFF
+		WHERE USERNAME='" . $username . "'
+		AND SYEAR='" . UserSyear() . "'
+		AND STAFF_ID!='" . UserStaffID() . "'
+		UNION SELECT 'exists'
+		FROM STUDENTS
+		WHERE USERNAME='" . $username . "'" );
+
+	if ( $existing_username )
+	{
+		$error[] = $username . ': ' . _( 'A user with that username already exists. Choose a different username and try again.' );
+
+		return 0;
+	}
+
+	$staff_id = DBSeqNextID( 'staff_staff_id_seq' );
+
+	if ( ! isset( $user['firstname'] ) )
+	{
+		// First and last name requires 'moodle/site:viewfullnames' capabitility.
+		$names = explode( ' ', $user['fullname'] );
+
+		$user['firstname'] = $names[0];
+
+		$user['lastname'] = $names[1];
+	}
+
+	if ( $profile == 'admin' )
+	{
+		$profile_id = '1';
+	}
+	elseif ( $profile == 'teacher' )
+	{
+		$profile_id = '2';
+	}
+	elseif ( $profile == 'parent' )
+	{
+		$profile_id = '3';
+	}
+
+	$sql = "INSERT INTO STAFF ";
+	$fields = 'SYEAR,STAFF_ID,LAST_NAME,FIRST_NAME,USERNAME,PROFILE,PROFILE_ID';
+	$values = "'" . UserSyear() . "','" . $staff_id . "','" . $user['firstname'] . "','" . $user['lastname'] . "','" .
+		$username . "','" . $profile . "','" . $profile_id . "'";
+
+	if ( ! empty( $user['email'] ) )
+	{
+		$fields .= ',EMAIL';
+		$values .= ",'" . $user['email'] . "'";
+	}
+
+	if ( ! empty( $_REQUEST['values']['PASSWORD_SET_USE_USERNAME'] ) )
+	{
+		$fields .= ',PASSWORD';
+		$values .= ",'" . encrypt_password( $username ) . "'";
+	}
+
+	$sql .= '(' . $fields . ') values(' . $values . ')';
+	DBQuery( $sql );
+
+	DBQuery( "INSERT INTO MOODLEXROSARIO (\"column\",rosario_id,moodle_id)
+		VALUES('staff_id','" . $staff_id . "'," . $user['id'] . ")" );
+
+	return $staff_id;
+}
+
+function ImportUsersFormConfirmCountdownJS( $class_prefix )
+{
+	?>
+	<script>
+	$(function(){
+		$('.<?php echo $class_prefix; ?>-form').submit(function(){
+
+			e.preventDefault();
+
+			var alertTxt = <?php echo json_encode(
+				_( 'Are you absolutely ready to import users? Make sure you have backed up your database!' )
+			); ?>;
+
+			// Alert.
+			if ( ! window.confirm( alertTxt ) ) return false;
+
+			var $buttons = $('.<?php echo $class_prefix; ?>-button'),
+				buttonTxt = $buttons.val(),
+				seconds = 5,
+				stopButtonHTML = <?php echo json_encode( SubmitButton(
+					_( 'Stop' ),
+					'',
+					'class="stop-button"'
+				) ); ?>;
+
+			$buttons.css('pointer-events', 'none').attr('disabled', true).val( buttonTxt + ' ... ' + seconds );
+
+			var countdown = setInterval( function(){
+				if ( seconds == 0 ) {
+					clearInterval( countdown );
+					$('.<?php echo $class_prefix; ?>-form').off('submit').submit();
+					return;
+				}
+
+				$buttons.val( buttonTxt + ' ... ' + --seconds );
+			}, 1000 );
+
+			// Insert stop button.
+			$( stopButtonHTML ).click( function(){
+				clearInterval( countdown );
+				$('.stop-button').remove();
+				$buttons.css('pointer-events', '').attr('disabled', false).val( buttonTxt );
+				return false;
+			}).insertAfter( $buttons );
+		});
+	});
+	</script>
+	<?php
+}