+++ CHANGES.md
@@ -1,6 +1,10 @@
 # CHANGES
 ## RosarioSIS Student Information System
 
+Changes in 5.9
+--------------
+- Import Moodle Users in Moodle/config.inc.php & Moodle/includes/ImportUsers.fnc.php
+
 Changes in 5.8
 --------------
 - Add `_listSearch()` function in ListOutput.fnc.php